+++ NEWS.md
@@ -1,4 +1,11 @@
+# SympluR 0.2.0
+
+* Supports access to the `twitter/analytics/tweets/activity` endpoint of the Symplur API.
+* Supports access to the `twitter/analytics/content/retweets` endpoint of the Symplur API.
+* Supports access to the `twitter/analytics/content/words` endpoint of the Symplur API.
+
 # SympluR 0.1.0
-- First release.
-- Supports access to the `twitter/analytics/tweets/summary` endpoint of the Symplur API.
-- Provides a process for looping through multiple databases to create a large data frame.
+
+* First release.
+* Supports access to the `twitter/analytics/tweets/summary` endpoint of the Symplur API.
+* Provides a process for looping through multiple databases to create a large data frame.